Got join games working! And multiple logins!

+++ shared/src/main/java/model/CreateResponse.java
@@ -1,4 +1,4 @@
 package model;
 
-public record CreateResponse(Integer gameID) {
+public record CreateResponse(Double gameID) {
 }
